+++ e2e/tests/steps/websocket.rs
@@ -1,11 +1,15 @@
+use std::time::Duration;
+
 use cucumber::{then, when};
+use tokio::time::sleep;
 
 use crate::World;
 
 #[when(regex = r"^(\S+) loses WS connection$")]
 async fn ws_connection_loss(world: &mut World, id: String) {
     let member = world.get_member(&id).unwrap();
-    member.ws_mock().enable_connection_loss(9999).await;
+    member.ws_mock().enable_connection_loss(3999).await;
+    sleep(Duration::from_millis(500)).await;
 }
 
 #[when(regex = r"^(\S+) restores WS connection$")]