+++ src/media/track/remote.rs
@@ -153,8 +153,9 @@ impl Track {
     /// Stops this [`Track`] invoking an `on_stopped` callback if it's in a
     /// [`MediaStreamTrackState::Live`] state.
     pub async fn stop(self) {
-        if self.0.track.ready_state().await == MediaStreamTrackState::Live {
-            self.0.track.stop().await;
+        let state = self.0.track.ready_state().await;
+        self.0.track.stop().await;
+        if state == MediaStreamTrackState::Live {
             self.0.on_stopped.call0();
         }
     }