+++ CHANGELOG.md
@@ -6,6 +6,20 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.4.1] · 2023-??-??
+[0.4.1]: /../../tree/medea-jason-0.4.1
+
+[Diff](https://github.com/instrumentisto/medea-jason/compare/medea-jason-0.4.0...medea-jason-0.4.1)
+
+### Fixed
+
+- Screen sharing in Firefox ([#135]).
+
+[#135]: /../../pull/135
+
+
+
+
 ## [0.4.0] · 2023-07-11
 [0.4.0]: /../../tree/medea-jason-0.4.0